refactor(config): move validation next to opts definition
